requirements-py312.txt
# ============================================================================
# Python 3.12 兼容的依赖包列表（用于 Paradex 虚拟环境）
# ============================================================================

# ────────────────────────────────────────────────────────────────────────────
# 🏛️ 核心框架 (Core Framework)
# ────────────────────────────────────────────────────────────────────────────
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.0
pydantic-settings==2.1.0
injector==0.21.0

# ────────────────────────────────────────────────────────────────────────────
# 🔄 异步和网络 (Async & Networking)
# ────────────────────────────────────────────────────────────────────────────
aiohttp>=3.9.0
aiofiles>=23.0.0
websockets>=12.0
websocket-client>=1.6.0
tenacity>=8.2.3  # EdgeX 重试机制

# ────────────────────────────────────────────────────────────────────────────
# 🔗 交易所适配器 (Exchange Adapters)
# ────────────────────────────────────────────────────────────────────────────
ccxt>=4.4.30
paradex-py  # Paradex 官方 SDK

# EdgeX 官方 SDK（从 GitHub 安装，导入 edgex_sdk）
# 安装命令: pip install git+https://github.com/edgex-Tech/edgex-python-sdk.git
git+https://github.com/edgex-Tech/edgex-python-sdk.git

# Lighter 官方 SDK（使用 --no-deps 安装以避免依赖冲突）
# 安装命令: pip install --no-deps git+https://github.com/elliottech/lighter-python.git
# 注意: lighter-sdk 需要 eth-account>=0.13.4，但 paradex-py 需要 eth-account<0.12.0
# 使用 --no-deps 可以绕过依赖检查，实际运行时 lighter SDK 在 eth-account 0.11.3 上可正常工作

# Hyperliquid SDK
hyperliquid-python-sdk
httpx>=0.25.0

# 签名依赖（版本兼容 Python 3.12）
eth-account>=0.10.0
eth-typing>=3.5.0
eth-utils>=2.3.0

# Backpack 签名依赖（ED25519 签名）
PyNaCl>=1.5.0  # Backpack 交易所 ED25519 签名库

# ────────────────────────────────────────────────────────────────────────────
# 📊 数据处理 (Data Processing) - Python 3.12 兼容版本
# ────────────────────────────────────────────────────────────────────────────
pandas>=2.1.0  # 最新版本支持 Python 3.12
numpy>=1.26.0  # Python 3.12 需要 numpy >= 1.26
plotly>=5.18.0
streamlit>=1.28.0

# ────────────────────────────────────────────────────────────────────────────
# 🖥️ 终端 UI (Terminal UI)
# ────────────────────────────────────────────────────────────────────────────
rich==13.7.0
psutil==5.9.6

# ────────────────────────────────────────────────────────────────────────────
# ⚙️ 配置管理 (Configuration)
# ────────────────────────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.0

# ────────────────────────────────────────────────────────────────────────────
# 🗄️ 基础设施 (Infrastructure)
# ────────────────────────────────────────────────────────────────────────────
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite>=0.19.0

# ────────────────────────────────────────────────────────────────────────────
# 🧪 测试框架 (Testing)
# ────────────────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

